backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
pydantic[email]
pytest
httpx
pgvector
sentence-transformers